move formatting into utils
