+++ src/main.rs
@@ -26,7 +26,6 @@ mod boards_scene;
 mod button_group;
 mod components;
 mod delete_scene;
-mod formatting;
 mod info_scene;
 mod input;
 mod list_scene;
@@ -42,7 +41,6 @@ use boards_scene::BoardInfo;
 use button_group::ButtonGroup;
 use components::*;
 use firefly_rust::*;
-use formatting::*;
 use input::{Input, InputManager};
 use scores_scene::ScoreInfo;
 use state::*;
